SP results

+++ bandwidthplot.py
@@ -17,7 +17,8 @@ labeltype = ['Intel Cascade Lake','NEC SX-Aurora 20B','NVIDIA A100 40GB']
 # it matches the order of instrs, ['SCEXAO','MICADO [1]','MICADO [2]','MAVIS','MAORY','EPICS']
 mnlist = [[2000,10000],[5000,10000],[5000,25000],[5000,20000],[8000,80000],[35000,70000]]  
 searchfolder = 'log'
-precision = 'double'
+#precision = 'double'
+precision = 'single'
 #exptypes = ['amd','cascadelake','NEC_20B','a100']
 exptypes = ['cascadelake','NEC_20B','a100']
